requirements.txt
attrs==19.3.0
Automat==20.2.0
baidu-aip==2.2.18.0
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
constantly==15.1.0
cryptography==2.9.2
cssselect==1.1.0
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
loguru==0.5.0
lxml==4.5.1
nose==1.3.7
parsel==1.6.0
peewee==3.13.3
Pillow==7.1.2
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
PyHamcrest==2.0.2
PyMySQL==0.9.3
pyOpenSSL==19.1.0
PySocks==1.7.1
pytesseract==0.3.4
queuelib==1.5.0
redis==3.5.3
requests==2.23.0
Scrapy==2.1.0
service-identity==18.1.0
six==1.15.0
soupsieve==2.0.1
stem==1.8.0
tesserocr==2.4.0
Twisted==20.3.0
urllib3==1.25.9
w3lib==1.22.0
win32-setctime==1.0.1
zope.interface==5.1.0
